in  this commit: - reworking helper functions structure - reworking helper functions names - support 5.3
