--- library/Barberry/file/helper.php
@@ -1,12 +1,0 @@
-<?php
-/**
- * @author lion
- */
-
-namespace Barberry\file {
-
-    function als($path)
-    {
-        return rtrim($path, DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR;
-    }
-}
